Add files via upload
This objective of this project is to encode and decode messages using a common key. This project will be built using the Tkinter and base64 library.

In this project, users have to enter the message to encode or decode. Users have to select the mode to choose the encoding and decoding process. The same key must be used to process the encoding and decoding for the same message.